+++ Race Element.Data/Common/AbstractSimulatorDataJob.cs
@@ -1,4 +1,4 @@
-﻿using RaceElement.Core.Jobs.LoopJob;
+﻿using RaceElement.Core.Jobs.Loop;
 using System;
 using System.Collections.Generic;
 using System.Linq;
