+++ Race Element.Data/Games/GameManager.cs
@@ -1,4 +1,4 @@
-﻿using RaceElement.Core.Jobs.LoopJob;
+﻿using RaceElement.Core.Jobs.Loop;
 using RaceElement.Data.Common;
 using RaceElement.Data.Games.iRacing;
