+++ Race Element.HUD.Common/Overlays/Driving/GForceTrace/GForceDataJob.cs
@@ -1,4 +1,4 @@
-﻿using RaceElement.Core.Jobs.LoopJob;
+﻿using RaceElement.Core.Jobs.Loop;
 using RaceElement.Data.Common;
 using RaceElement.Util.SystemExtensions;
 using System.Numerics;
