+++ Race_Element.HUD.ACC/Overlays/Driving/RainPrediction/RainPredictionJob.cs
@@ -1,4 +1,4 @@
-﻿using RaceElement.Core.Jobs.LoopJob;
+﻿using RaceElement.Core.Jobs.Loop;
 using System;
 using System.Collections.Concurrent;
 using System.Diagnostics;
