+++ Race_Element.HUD.ACC/Overlays/Pitwall/OverlayDualSenseX/DualSenseXJob.cs
@@ -1,4 +1,4 @@
-﻿using RaceElement.Core.Jobs.LoopJob;
+﻿using RaceElement.Core.Jobs.Loop;
 using System;
 using static RaceElement.HUD.ACC.Overlays.Pitwall.OverlayDualSenseX.DualSenseXResources;
